Moved file utility functions to a new file
