+++ src/notifications.ts
@@ -1,4 +1,5 @@
 import { Notification } from "./store/definitions";
+import { shortenFileName } from "./utils/file";
 
 /*
  * const exampleNotification: Notification = {
@@ -8,21 +9,6 @@ import { Notification } from "./store/definitions";
  *     type: "neutral"
  * };
  */
-const MAX_DISPLAYED_CHAR_FOR_FIRST_PART = 14;
-const MAX_DISPLAYED_CHAR_FOR_SECOND_PART = 5;
-
-const shortenFileName = (name: string, maxChar = 20) => {
-    const [, filename, fileExtension] = name.match(/^(.+)(\.[\S]*)$/) || [];
-    if (!filename) {
-        return name;
-    }
-
-    if (filename.length <= maxChar) return name;
-
-    return `${filename.substring(0, MAX_DISPLAYED_CHAR_FOR_FIRST_PART)}[...]${filename.substring(
-        filename.length - MAX_DISPLAYED_CHAR_FOR_SECOND_PART
-    )}${fileExtension}`;
-};
 
 const fileAttachmentAlreadyAttachedNotification = ({ fileName }: { fileName: string }): Notification => ({
     id: `FileAttachmentAlreadyAttachedNotification_${Math.random()}`,
@@ -103,6 +89,5 @@ export const notifications = {
     fileDownloadInvalidSizeNotification,
     fileDownloadInvalidTypeNotification,
     noConnectionNotification,
-    failedToInitSessionNotification,
-    shortenFileName
+    failedToInitSessionNotification
 };
